refactor(bench): consolidate cold-start + regression-check into tests/benchmark.ts
Inlines tests/cold-start-benchmark.ts (297 LOC) and
tests/perf-regression-check.ts (369 LOC) into the existing
tests/benchmark.ts as flag-dispatched modes:

  npx tsx tests/benchmark.ts                # executor bench (default)
  npx tsx tests/benchmark.ts --cold-start   # spawn-to-MCP-ready latency
  npx tsx tests/benchmark.ts --check        # vs perf-baseline.json
  npx tsx tests/benchmark.ts --update       # overwrite platform baseline
  npx tsx tests/benchmark.ts --self-test    # threshold + extract assertions

- Single dispatcher routes by argv flags before any mode runs.
- Symbols renamed where they collided: main → executorMain,
  fmt → fmtMs/fmtUnit, printTable kept for executor / printRegressionTable
  for delta rows, check → checkRegressions, update → updateBaseline.
- runJsonBench now accepts extraArgs so regression mode invokes
  tests/benchmark.ts --cold-start --json and tests/benchmark.ts --json
  instead of the deleted standalone scripts.
- package.json bench:* scripts and .github/workflows/bench.yml updated
  to point at the consolidated entrypoint.
- Self-test passes 13 assertions (threshold floor, regression/improved/
  ok/new classification, extractMetrics filtering).

No new test files in tests/. Behavior unchanged — only file layout.
