+++ package.json
@@ -96,10 +96,10 @@
     "test": "vitest run",
     "test:watch": "vitest",
     "benchmark": "npx tsx tests/benchmark.ts",
-    "bench:cold-start": "npx tsx tests/cold-start-benchmark.ts",
-    "bench:baseline:check": "npx tsx tests/perf-regression-check.ts --check",
-    "bench:baseline:update": "npx tsx tests/perf-regression-check.ts --update",
-    "bench:baseline:self-test": "npx tsx tests/perf-regression-check.ts --self-test",
+    "bench:cold-start": "npx tsx tests/benchmark.ts --cold-start",
+    "bench:baseline:check": "npx tsx tests/benchmark.ts --check",
+    "bench:baseline:update": "npx tsx tests/benchmark.ts --update",
+    "bench:baseline:self-test": "npx tsx tests/benchmark.ts --self-test",
     "test:use-cases": "npx tsx tests/use-cases.ts",
     "test:compare": "npx tsx tests/context-comparison.ts",
     "test:ecosystem": "npx tsx tests/ecosystem-benchmark.ts",
